Stable release v1.0
- Modify README.md and add full description of program
- Fix code in 'sentiment_analyser_(FastText).py'
- Rename 'sentiment_analyser_(FastText).py' on 'sentiment_analyser.py'
